requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
passlib[bcrypt]
PyJWT
requests
langchain
langchain-groq
pypdf
gunicorn